+++ src/lib_reload_events.rs
@@ -189,7 +189,12 @@ impl LibReloadNotifier {
             let n = subscribers.len();
             log::trace!("sending {evt:?} to {n} subscribers");
             // keep only those subscribers that are still around and kicking.
-            subscribers.retain(|tx| tx.send(evt.clone()).is_ok());
+            subscribers.retain(|tx| {
+                if let Err(err) = tx.send(evt.clone()) {
+                    log::error!("Error sending event to subscriber: {err}");
+                }
+                true
+            });
             let removed = n - subscribers.len();
             if removed > 0 {
                 log::debug!(